- modified:   src/modules/rlm_sql/conf.h - modified:   src/modules/rlm_sql/rlm_sql.c
Added rlm_sql_post_proxy() function, so that one can use the rlm_sql module
to look up additional attributes for a user, and add it to the reply
after the request has been handled by the home-server. The table may be
the same as the authreply_table or a similar one.

- modified:   raddb/sites-available/default
- modified:   raddb/sql.conf
- modified:   raddb/sql/mssql/dialup.conf
- modified:   raddb/sql/mysql/dialup.conf
- modified:   raddb/sql/oracle/dialup.conf
- modified:   raddb/sql/postgresql/dialup.conf

Configuration samples for rlm_sql_post_proxy